brand full, but without test

+++ src/main/java/com/licoreria/licoreria/brands/domain/BrandsRepository.java
@@ -1,11 +1,5 @@
 package com.licoreria.licoreria.brands.domain;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-import java.math.BigDecimal;
 
-public interface BrandsRepository extends JpaRepository<Brand, BigDecimal> {
-
-
-
+public interface BrandsRepository extends JpaRepository<Brand, Long> {
 }
